reorganize page components
